Feat/16/be fe rest api comunication

+++ backend/library/src/main/java/com/hae/library/service/BookService.java
@@ -2,6 +2,7 @@ package com.hae.library.service;
 
 import com.hae.library.domain.Book;
 import com.hae.library.domain.BookInfo;
+import com.hae.library.domain.Category;
 import com.hae.library.domain.Enum.BookStatus;
 import com.hae.library.dto.Book.RequestBookWithBookInfoDto;
 import com.hae.library.dto.Book.ResponseBookWithBookInfoDto;
@@ -9,8 +10,10 @@ import com.hae.library.dto.BookInfo.ResponseBookInfoDto;
 import com.hae.library.dto.BookInfo.ResponseBookInfoWithBookDto;
 import com.hae.library.global.Exception.errorCode.BookErrorCode;
 import com.hae.library.global.Exception.RestApiException;
+import com.hae.library.global.Exception.errorCode.CategoryErrorCode;
 import com.hae.library.repository.BookInfoRepository;
 import com.hae.library.repository.BookRepository;
+import com.hae.library.repository.CategoryRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -30,11 +33,12 @@ public class BookService {
     private final BookRepository bookRepo;
     private final BookInfoRepository bookInfoRepo;
     private final BookInfoService bookInfoService;
+    private final CategoryRepository categoryRepo;
 
     // 새로운 책을 생성하는 메서드입니다.
     @Transactional
     public ResponseBookWithBookInfoDto createBook(RequestBookWithBookInfoDto requestBookWithBookInfoDto) {
-        log.error("requestBookDto: {}", requestBookWithBookInfoDto);
+        log.error("requestBookDto???: {}", requestBookWithBookInfoDto);
 
         // 청구기호가 중복되는지 확인합니다. 중복되는 경우 예외 처리합니다.
         if (bookRepo.existsByCallSign(requestBookWithBookInfoDto.getCallSign())) {
@@ -47,27 +51,18 @@ public class BookService {
         ResponseBookWithBookInfoDto responseBookWithBookInfoDto;
         // BookInfo가 존재하는 경우와 존재하지 않는 경우 다르게 처리합니다.
         if (bookInfoOptional.isPresent()) {
-            // BookInfo가 존재하는 경우
+            // BookInfo가 존재하는 경우 기존 BookInfo를 사용합니다.
             BookInfo bookInfo = bookInfoOptional.get();
+
             responseBookWithBookInfoDto =
                     saveBookWithBookInfo(requestBookWithBookInfoDto,
                             bookInfo);
         } else {
-            // BookInfo가 존재하지 않는 경우
-            // 새로운 BookInfo를 생성합니다.
-            ResponseBookInfoDto responseBookInfo = bookInfoService.createBookInfo(requestBookWithBookInfoDto);
-            BookInfo bookInfo = BookInfo.builder()
-                    .id(responseBookInfo.getId())
-                    .title(responseBookInfo.getTitle())
-                    .author(responseBookInfo.getAuthor())
-                    .isbn(responseBookInfo.getIsbn())
-                    .image(responseBookInfo.getImage())
-                    .publisher(responseBookInfo.getPublisher())
-                    .publishedAt(responseBookInfo.getPublishedAt())
-                    .build();
+            // BookInfo가 존재하지 않는 경우 새로운 BookInfo를 생성합니다.
+            BookInfo newBookInfo = bookInfoService.createBookInfo(requestBookWithBookInfoDto);
             // 새로 생성한 BookInfo를 사용하여 책을 저장합니다.
             responseBookWithBookInfoDto =
-                    saveBookWithBookInfo(requestBookWithBookInfoDto, bookInfo);
+                    saveBookWithBookInfo(requestBookWithBookInfoDto, newBookInfo);
         }
         return responseBookWithBookInfoDto;
     }
@@ -83,6 +78,7 @@ public class BookService {
         // 생성한 책 객체에 BookInfo를 추가합니다.
         book.addBookInfo(bookInfo);
         // 책 객체를 데이터베이스에 저장합니다.
+
         Book updateBook = bookRepo.save(book);
         return ResponseBookWithBookInfoDto.from(updateBook);
     }
@@ -135,14 +131,23 @@ public class BookService {
         Book book = bookRepo.findById(requestBookWithBookInfoDto.getId()).orElseThrow(() -> new RestApiException(BookErrorCode.BAD_REQUEST_BOOK));
         BookInfo bookInfo = book.getBookInfo();
 
-        // TODO: 청구기호 저장하기전에 중복되는지 확인
-        if (bookRepo.existsByCallSign(requestBookWithBookInfoDto.getCallSign())) {
+        // 청구기호 저장하기전에 중복되는지 확인합니다.
+        if (bookRepo.existsByCallSignAndIdIsNot(requestBookWithBookInfoDto.getCallSign(),
+                book.getId())) {
             throw new RestApiException(BookErrorCode.DUPLICATE_CALLSIGN);
         }
 
+        // 카테고리를 조회하고 카테고리가 존재하지 않는다면 예외를 발생시킵니다.
+        Category category =
+                categoryRepo.findByCategoryName(requestBookWithBookInfoDto.getCategoryName()).orElseThrow(() -> new RestApiException(CategoryErrorCode.BAD_REQUEST_CATEGORY));
+
+
         // 도서 정보를 업데이트하고, 업데이트된 도서 정보를 데이터베이스에 저장합니다.
-        bookInfo.updateBookInfo(requestBookWithBookInfoDto.getTitle(), requestBookWithBookInfoDto.getIsbn(), requestBookWithBookInfoDto.getAuthor(),
-                requestBookWithBookInfoDto.getPublisher(), requestBookWithBookInfoDto.getPublishedAt(), requestBookWithBookInfoDto.getImage());
+        bookInfo.updateBookInfo(requestBookWithBookInfoDto.getTitle(),
+                requestBookWithBookInfoDto.getIsbn(), requestBookWithBookInfoDto.getAuthor(),
+                requestBookWithBookInfoDto.getPublisher(),
+                requestBookWithBookInfoDto.getPublishedAt(),
+                requestBookWithBookInfoDto.getImage(), category);
         bookInfoRepo.save(bookInfo);
 
         // 청구기호와 도서의 상태 그리고 기증자 정보를 업데이트하고, 업데이트된 도서를 데이터베이스에 저장합니다.
@@ -158,11 +163,19 @@ public class BookService {
         // 삭제할 도서를 ID를 이용하여 검색합니다. 도서를 찾지 못하면 예외를 발생시킵니다.
         Book book = bookRepo.findById(bookId).orElseThrow(() -> new RestApiException(BookErrorCode.BAD_REQUEST_BOOK));
 
-        // TODO: 도서가 대여중인지 확인
-        // TODO: 마지막 도서이면 BookInfo도 삭제
+        // TODO: 도서가 대여중인지 확인합니다.
+        if (book.isLendingStatus() == true) {
+            throw new RestApiException(BookErrorCode.NOT_DELETE_BECAUSE_RENT_BOOK);
+        }
+
+        // TODO: 마지막 도서이면 BookInfo도 삭제합니다.
+        if (bookRepo.countByBookInfo(book.getBookInfo()) == 1) {
+            bookInfoRepo.delete(book.getBookInfo());
+        }
 
         // 찾은 도서를 데이터베이스에서 삭제합니다.
         bookRepo.deleteById(bookId);
+
     }
 
 }
